chnaged type nonce

+++ gameplay/queue/chore.go
@@ -349,8 +349,7 @@ func (chore *Chore) FinishWithWinResult(ctx context.Context, winResult WinResult
 			}
 		}
 
-		// TODO: int64 > int if 32bit os
-		if winResult.GameResult.Transaction, err = chore.currencywaitlist.Create(ctx, user.ID, *winResult.Value, int64(request.Nonce)); err != nil {
+		if winResult.GameResult.Transaction, err = chore.currencywaitlist.Create(ctx, user.ID, *winResult.Value, request.Nonce); err != nil {
 			chore.log.Error("could not create item of currencywaitlist", ChoreError.Wrap(err))
 			return
 		}
